assistant/requirements.txt
asgiref==3.5.2
asttokens==2.0.8
backcall==0.2.0
beautifulsoup4==4.11.1
certifi==2022.9.24
charset-normalizer==2.1.1
decorator==5.1.1
Django==4.1.1
djangorestframework==3.14.0
executing==1.1.0
idna==3.4
ipython==8.5.0
jedi==0.18.1
lxml==4.9.1
matplotlib-inline==0.1.6
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
prompt-toolkit==3.0.31
psycopg2==2.9.4
ptyprocess==0.7.0
pure-eval==0.2.2
Pygments==2.13.0
python-dotenv==0.21.0
pytz==2022.2.1
requests==2.28.1
six==1.16.0
soupsieve==2.3.2.post1
sqlparse==0.4.3
stack-data==0.5.1
traitlets==5.4.0
urllib3==1.26.12
wcwidth==0.2.5
